refactor(ping): delete tree.go + graph.go, rename tree2→tree, add --testenv
Cleans up the ping/ subdir to a single tree implementation —
the bubbletea-based scrollable TUI that was `ping tree2`. The
old `ping tree` and `ping graph` commands are removed; only what
operators actually need from them lands in the surviving file.

What's gone:

  graph.go (4528 lines) — `ping graph` command. Self-described as
    deprecated in its own help text ("Consider using 'ping tree'
    instead"). All output via fmt.Println which causes the terminal-
    jump-on-write scroll problem the operator complained about
    (you can't scroll up to read earlier results while it's still
    running). Three monstrous functions (init=845, runTreeViewMode
    =2838, runDmsgTreeViewMode=741) all duplicated by the cleaner
    tree2 implementation.

  tree.go (430 lines) — `ping tree` command. Thin wrapper that
    re-declared 31 of its own flags (treeXxx), copied them into
    graph.go's globals (graphXxx), and called graph.go's view
    functions. Adds no functionality the surviving tree2 didn't
    already have.

What moved instead of deleted:

  --testenv flag + isTestEnv() + getDeployment() — preserved from
    the old tree.go and brought to the surviving tree command;
    useful for running against the test deployment without
    exporting SKYWIRETEST=1.

  transportEntry / uptimeEntry / treeNeighbor types — moved from
    graph.go to a new tree_helpers.go since the surviving tree
    command still uses them for TPD JSON parsing and adjacency
    map BFS.

What's noted in comments for future reference:

  - Cache directory vs cache file. Old `ping tree` flagged
    cache locations as directories (--cdt /tmp/tpd) with
    per-endpoint files synthesized inside. Surviving `ping tree`
    flags as individual files (--cft /tmp/tpd.json). The
    directory model is more flexible if a caller needs multiple
    distinct caches under one host — comment in tree_helpers.go
    notes git history as the path back to the helpers if needed.

The cobra `Use: "tree2"` is renamed to `Use: "tree"` and the file
tree2.go is renamed to tree.go via `git mv`. The renamed-from-tree2
ping command is now what shows up as `skywire cli visor ping tree`.

root.go's RootCmd.Long subcommand listing dropped the now-deleted
"ping tree", "ping tree2", "ping graph" entries — the surviving
list is `ping <pk>`, `ping test`, `ping tree`, `ping stop-all`.

No behavior change for the surviving tree command beyond:
  - It's reachable at `ping tree` now (was `ping tree2`)
  - It accepts `--testenv` (didn't before)

Diff stat: ~5000 lines deleted across two files; ~130 lines added
in tree_helpers.go for the preserved helpers + types.
